add possibility to define default options via hiera
if the default options are not set (undef) the old behaviour
is used by merging the defaults from params.pp.
If the default options are set (eg. via hiera or as parameter),
the params.pp settings for the options is ignored for merging.

This let us make a transition to hiera defined defaults.